+++ dockstore-webservice/src/main/java/io/dockstore/webservice/core/LicenseInformation.java
@@ -1,14 +1,10 @@
 package io.dockstore.webservice.core;
 
-import javax.persistence.Column;
 import javax.persistence.Embeddable;
 
 @Embeddable
 public class LicenseInformation {
     String licenseName;
-
-    @Column(columnDefinition = "text")
-    String licenseContent;
 
     public String getLicenseName() {
         return licenseName;
@@ -16,13 +12,5 @@ public class LicenseInformation {
 
     public void setLicenseName(String licenseName) {
         this.licenseName = licenseName;
-    }
-
-    public String getLicenseContent() {
-        return licenseContent;
-    }
-
-    public void setLicenseContent(String licenseContent) {
-        this.licenseContent = licenseContent;
     }
 }
